2016 Update

+++ Task/Factorial/Groovy/factorial-1.groovy
@@ -1,2 +1,2 @@
 def rFact
-rFact = { (it > 1) ? it * rFact(it - 1) : 1 }
+rFact = { (it > 1) ? it * rFact(it - 1) : 1 as BigInteger }
